Enum gedcomx_date::GedcomxDate

pub enum GedcomxDate {
    Simple(Simple),
    Range(Range),
    Recurring(Recurring),
}

Gedcomx date Enum that holds the three types of gedcomx dates

Variants

Simple

Simple date. See 5.2 Simple date, 5.7 Approximate date.

Range

Date range. See 5.4 Closed date range, 5.5 Open ended date range, 5.8 Approximate date range.

Recurring

Recurring date. See 5.6 Recurring date.

Methods

impl GedcomxDate

GedxomxDate implementation. You can match over the parse result, or use these methods.

fn get_simple_date(&self) -> Option<Simple>

fn get_range(&self) -> Option<Range>

fn get_recurring(&self) -> Option<Recurring>